% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_epidemic.R
\name{is_epidemic}
\alias{is_epidemic}
\title{Epidemic Comptability}
\usage{
is_epidemic(r, i)
}
\arguments{
\item{r}{numeric vector of removal times}

\item{i}{numeric vector of infection times}
}
\value{
boolean
}
\description{
Indicate if infection and removal times are compatible.
}
