% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_beta.R
\name{map_beta}
\alias{map_beta}
\title{Assign Infection Rates}
\usage{
map_beta(beta, map)
}
\arguments{
\item{beta}{numeric vector of infection rates}

\item{map}{matrix of assignments}
}
\value{
matrix of infection rates
}
\description{
Assign individual infection rates.
}
