% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_gsem.R
\name{mle_gsem}
\alias{mle_gsem}
\title{MLE for Completely Observed SEM}
\usage{
mle_gsem(r, i, N)
}
\arguments{
\item{r}{numeric vector of removal times}

\item{i}{numeric vector of infection times}

\item{N}{integer population size}
}
\value{
MLE for (beta, gamma)
}
\description{
Compute MLE for completely observed stochastic epidemic.
}
