% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multitypes.R
\name{multitypes}
\alias{multitypes}
\title{Assign Infection Rates}
\usage{
multitypes(beta, map)
}
\arguments{
\item{beta}{numeric vector of infection rates}

\item{map}{matrix of assignments}
}
\value{
matrix of infection rates
}
\description{
Assign individual infection rates.
}
