% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncda_mcmc_sem.R
\name{ncda_mcmc_sem}
\alias{ncda_mcmc_sem}
\title{Non-Centered Data-Augmented MCMC Sampler for SEM}
\usage{
ncda_mcmc_sem(
  niter,
  sd,
  J,
  r,
  N,
  m,
  betastart,
  betashape,
  betarate,
  gammastart,
  gammashape,
  gammarate
)
}
\arguments{
\item{niter}{integer iterations}

\item{sd}{positive control of random walk proposal}

\item{J}{integer infection time updates}

\item{r}{numeric vector of removal times}

\item{N}{integer population size}

\item{m}{integer Erlang shape}

\item{betastart}{positive}

\item{betashape}{prior}

\item{betarate}{prior}

\item{gammastart}{positive}

\item{gammashape}{prior}

\item{gammarate}{prior}
}
\value{
array
}
\description{
Draw posterior samples for a stochastic epidemic model.
}
