% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbla_ed_homo.R
\name{pbla_ed_homo}
\alias{pbla_ed_homo}
\title{Eichner-Dietz PBLA (Homogeneous Mixing)}
\usage{
pbla_ed_homo(r, beta, gamma, N, m = 1, nt = 100, mint = -2)
}
\arguments{
\item{r}{numeric vector of increasing removal times}

\item{beta}{numeric rate}

\item{gamma}{numeric rate}

\item{m}{positive integer shape}

\item{nt}{integer points for trapezoidal integration}

\item{mint}{numeric lower bound for trapezoidal integration}
}
\value{
negative log likelihood
}
\description{
Compute the Eichner-Dietz likelihood approximation. Supports Erlang infectious periods.
}
