% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbla_exp.R
\name{pbla_exp}
\alias{pbla_exp}
\title{Standard PBLA (Exponential)}
\usage{
pbla_exp(r, beta, gamma, lag = 0)
}
\arguments{
\item{r}{numeric vector of increasing removal times}

\item{beta}{matrix of rates}

\item{gamma}{numeric vector of rates}

\item{lag}{numeric fixed lag}
}
\value{
negative log likelihood
}
\description{
Compute pair-based likelihood approximation. Assume exponential infectious periods.
}
