% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbla_f.R
\name{pbla_f}
\alias{pbla_f}
\title{f-based PBLA}
\usage{
pbla_f(r, beta, gamma, lag = 0)
}
\arguments{
\item{r}{numeric vector of removal times}

\item{beta}{matrix of rates}

\item{gamma}{numeric vector of rates}

\item{lag}{numeric fixed lag}
}
\value{
negative log likelihood
}
\description{
Via 3.3.1., compute pair-based likelihood approximation. Supports exponential infectious periods.
}
