% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbla_gsem.R
\name{pbla_gsem}
\alias{pbla_gsem}
\title{PBLAs for General Stochastic Epidemic Model}
\usage{
pbla_gsem(rates, pbla, r, N, etc = NULL)
}
\arguments{
\item{rates}{numeric vector of rates}

\item{pbla}{function}

\item{r}{numeric vector of increasing removal times}

\item{N}{integer population size}

\item{etc}{other parameters to pass (e.g. lag)}
}
\value{
negative log likelihood
}
\description{
Run a pair-based likelihood approximation for a general stochastic epidemic model. Compatible with \code{pbla_prod}, \code{pbla_weak}, \code{pbla_std_gsem}, \code{pbla_sep_gsem}, \code{pbla_f_gsem}, and \code{pbla_ed_gsem}.
}
