% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbla_het.R
\name{pbla_het}
\alias{pbla_het}
\title{PBLAs for Stochastic Epidemic Model (Heterogeneous Mixing)}
\usage{
pbla_het(rates, R, pbla, imap, rmap, r, etc)
}
\arguments{
\item{rates}{numeric vector of rates}

\item{R}{integer index to split rates on}

\item{pbla}{function}

\item{imap}{matrix of assignments}

\item{rmap}{vector of assignments}

\item{r}{numeric vector of increasing removal times}

\item{etc}{other parameters to pass (e.g. lag)}
}
\value{
negative log likelihood
}
\description{
Run a pair-based likelihood approximation for a stochastic epidemic model. Heterogeneous mixing available via mapping assignments.
}
