% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbla_homo.R
\name{pbla_homo}
\alias{pbla_homo}
\title{PBLAs for Stochastic Epidemic Model (Homogeneous Mixing)}
\usage{
pbla_homo(rates, pbla, r, N, etc)
}
\arguments{
\item{rates}{numeric vector of rates}

\item{pbla}{function}

\item{r}{numeric vector of increasing removal times}

\item{N}{integer population size}

\item{etc}{other parameters to pass (e.g. lag)}
}
\value{
negative log likelihood
}
\description{
Run a pair-based likelihood approximation for a stochastic epidemic model. Heterogeneous mixing available via mapping assignments.
}
