% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbla_mcmc.R
\name{pbla_mcmc}
\alias{pbla_mcmc}
\title{PBLA MCMC for Stochastic Epidemic Models}
\usage{
pbla_mcmc(
  niter,
  sd,
  pbla,
  betamap,
  gammamap,
  r,
  etc,
  betastart,
  betashape,
  betarate,
  gammastart,
  gammashape,
  gammarate
)
}
\description{
Use PBLA in MCMC with Metropolis-Hastings-based random walk.
}
\details{
@param niter integer iterations
@param sd standard deviation of random walk proposal
@param pbla likelihood approximation
@param betamap matrix of assignments
@param gammamap vector of assignments
@param r numeric vector of increasing removal times
@param etc other parameters to pass (e.g. lag)
@param betastart positive initial value
@param betashape positive prior
@param betarate positive prior
@param gammastart positive initial value
@param gammashape positive prior
@param gammarate positive prior

@return posterior sample

@export
}
