% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbla_multi.R
\name{pbla_multi}
\alias{pbla_multi}
\title{PBLAs for Multitype Stochastic Epidemic Model}
\usage{
pbla_multi(rates, R, pbla, betamap, gammamap, r, etc = NULL)
}
\arguments{
\item{rates}{numeric vector of rates}

\item{R}{integer index to split rates on}

\item{pbla}{function}

\item{betamap}{matrix of assignments}

\item{gammamap}{vector of assignments}

\item{r}{numeric vector of increasing removal times}

\item{etc}{other parameters to pass (e.g. lag)}
}
\value{
negative log likelihood
}
\description{
Run a pair-based likelihood approximation for a multitype stochastic epidemic model. Multitypes available via mapping assignments. Compatible with \code{pbla_std}, \code{pbla_ed}, \code{pbla_sep}, and \code{pbla_f}.
}
