% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbla_sep.R
\name{pbla_sep}
\alias{pbla_sep}
\title{Separated Product PBLA}
\usage{
pbla_sep(r, beta, gamma, lag = 0)
}
\arguments{
\item{r}{numeric vector of increasing removal times}

\item{beta}{matrix of rates}

\item{gamma}{numeric vector of rates}

\item{lag}{numeric fixed lag}
}
\value{
negative log likelihood
}
\description{
Based on product independence, compute pair-based likelihood approximation. Assume exponential infectious periods.
}
