% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbla_sep_gsem.R
\name{pbla_sep_gsem}
\alias{pbla_sep_gsem}
\title{Separated Product PBLA (General SEM)}
\usage{
pbla_sep_gsem(r, beta, gamma, N, A = 1, lag = 0)
}
\arguments{
\item{r}{numeric vector of increasing removal times}

\item{beta}{numeric rate}

\item{gamma}{numeric rate}

\item{N}{integer population size}

\item{A}{integer patient zeros}

\item{lag}{numeric fixed lag}
}
\value{
negative log likelihood
}
\description{
Based on product independence, compute pair-based likelihood approximation. Supports exponential infectious periods.
}
