% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbla_std.R
\name{pbla_std}
\alias{pbla_std}
\title{Standard PBLA}
\usage{
pbla_std(r, beta, gamma, m = 1, lag = 0)
}
\arguments{
\item{r}{numeric vector of increasing removal times}

\item{beta}{matrix of rates}

\item{gamma}{numeric vector of rates}

\item{m}{positive integer shape}

\item{lag}{numeric fixed lag}
}
\value{
negative log likelihood
}
\description{
Compute pair-based likelihood approximation. Supports Erlang infectious periods.
}
