% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbla_std_gsem.R
\name{pbla_std_gsem}
\alias{pbla_std_gsem}
\title{Standard PBLA (General SEM)}
\usage{
pbla_std_gsem(r, beta, gamma, N, m = 1, A = 1, lag = 0)
}
\arguments{
\item{r}{numeric vector of increasing removal times}

\item{beta}{numeric rate}

\item{gamma}{numeric rate}

\item{N}{integer population size}

\item{m}{positive integer shape}

\item{A}{integer patient zeros}

\item{lag}{numeric fixed lag}
}
\value{
negative log likelihood
}
\description{
Compute pair-based likelihood approximation. Supports Erlang infectious periods.
}
