% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbla_weak.R
\name{pbla_weak}
\alias{pbla_weak}
\title{Weak limit PBLA}
\usage{
pbla_weak(r, beta, gamma, N, A = 1, lag = 0)
}
\arguments{
\item{r}{numeric vector of increasing removal times}

\item{beta}{numeric rate}

\item{gamma}{numeric rate}

\item{N}{integer population size}

\item{A}{integer patient zeros}

\item{lag}{numeric fixed lag}
}
\value{
negative log likelihood
}
\description{
Based on weak limit result, compute pair-based likelihood approximation. Assumes exponential infectious periods.
}
