% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerlsem.R
\name{rerlsem}
\alias{rerlsem}
\title{Simulate Erlang Infectious Periods}
\usage{
rerlsem(beta, gamma, m, N)
}
\arguments{
\item{beta}{numeric rate}

\item{gamma}{numeric rate}

\item{m}{integer shape}

\item{N}{integer population size}
}
\value{
infection and removal times
}
\description{
Draw Erlang infectious periods assuming SEM model.
}
