% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rexpsem.R
\name{rexpsem}
\alias{rexpsem}
\title{Simulate Exponential Infectious Periods}
\usage{
rexpsem(beta, gamma, N)
}
\arguments{
\item{beta}{numeric rate}

\item{gamma}{numeric rate}

\item{N}{integer population size}
}
\value{
infection and removal times
}
\description{
Draw exponential infectious periods assuming SEM model.
}
