% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgsem.R
\name{rgsem}
\alias{rgsem}
\title{Simulate Infectious Periods}
\usage{
rgsem(beta, gamma, N, m = 1)
}
\arguments{
\item{beta}{numeric rate}

\item{gamma}{numeric rate}

\item{N}{integer population size}

\item{m}{integer shape}
}
\value{
infection and removal times
}
\description{
Draw infectious periods for general stochastic epidemic.
}
