% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpseudor.R
\name{rpseudor}
\alias{rpseudor}
\title{Pseudo Removal Times}
\usage{
rpseudor(r, eta)
}
\arguments{
\item{r}{removal times}

\item{eta}{proportion reported}
}
\value{
removal times
}
\description{
Simulate new removal times from normal kernel density estimate.
}
