% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbla_ed.R
\name{pbla_ed}
\alias{pbla_ed}
\title{Eichner-Dietz PBLA}
\usage{
pbla_ed(r, beta, gamma, m, nt, mint)
}
\arguments{
\item{r}{numeric vector of increasing removal times}

\item{beta}{matrix of rates}

\item{gamma}{numeric vector of rates}

\item{m}{positive integer shape}

\item{nt}{integer points for trapezoidal integration}

\item{mint}{numeric lower bound for trapezoidal integration}
}
\value{
negative log likelihood
}
\description{
Compute the Eichner-Dietz likelihood approximation. Supports Erlang infectious periods.
}
